ensure that _foreColour and _backColour get set
I think this is what was intended in bcbdf3e. The other option
was to set them in `TheView.__init__`. But, I saw that changing
them to None in `__init__` was part of a bug fix back in 0c44f29.

Noticed `self.wrapper._backColour.set()` crashing first time
Waterfall was run without the defaults having been set, yet.